validate workchain inputs (#62)
Validates `parameters` input of all WorkChains using `voluptuous` schemas (both for the keys present and for the type of their values).

Validation happens when passing the builder to the WorkChain, i.e. the workchain is not submitted if the dictionary fails validation.

The behavior of the `pressure_list` and `temperature_list` keys is modified slightly: if you don't want to provide them, simply don't set the key (instead of setting the key to `None`).

The schema is part of the API documentation; users can `print(myworkchain.parameters_info)` to check which parameters the workchain needs.