catch exceptions when prediction oxidation states
The oximachine_runner would throw exceptions in a number of scenarios
(no metal in structure, parsing failed, featurization failed, ...).

This commit catches these exceptions and returns an empty list of
predictions in case an exception is encountered.

It also switches the `test.yaml` multistage protocol to use the
oximachine for the initial magnetization (before, it was the only
protocol using "element", which caused confusion).